Skip the git pull on bare repositories
A second run with -backup.bare-clone would fail as it would still try to
retrieve a work tree and do a git pull. Now it will check whether a
discovered repository is bare or not. And if it is, it will just skip
the git pull and only do a git fetch down the line.